+++ src/__test__/components/EditOntologyDeleteEdge.test.js
@@ -6,7 +6,7 @@ import EditOntologyDeleteEdge from '../../components/EditOntologyDeleteEdge'
 const setup = () => {
   const props = {
     type: 'edge',
-    opearation: 'delete',
+    operation: 'delete',
     setStoreState: jest.fn(),
     addToArray: jest.fn(),
     removeFromObject: jest.fn(),