+++ src/i18n/en.js
@@ -274,6 +274,26 @@ export default {
   toNode: 'To node',
   toAndFromNodesEqual: 'To and from nodes cannot be the same!',
   connectionAlreadyExists: 'Connection already exists',
+  couldNotAddNode: 'Could not add node',
+  couldNotUpdateNode: 'Could not update node',
+  couldNotDeleteNode: 'Could not delete node',
+  couldNotRestoreNode: 'Could not restore node',
+  couldNotAddEdge: 'Could not add edge',
+  couldNotUpdateEdge: 'Could not update edge',
+  couldNotDeleteEdge: 'Could not delete edge',
+  couldNotRestoreEdge: 'Could not restore edge',
+  nodeAdded: 'Node added',
+  nodeUpdated: 'Node updated',
+  nodeDeleted: 'Node deleted',
+  nodeRestored: 'Node restored',
+  edgeAdded: 'Edge added',
+  edgeUpdated: 'Edge updated',
+  edgeDeleted: 'Edge deleted',
+  edgeRestored: 'Edge restored',
+  nodesDeleted: 'Nodes deleted',
+  nodesRestored: 'Nodes restored',
+  edgesDeleted: 'Edges deleted',
+  edgesRestored: 'Edges restored',
 
   // footer navbar
   hidePhysicsSettings: 'Hide physics settings',
@@ -347,4 +367,8 @@ export default {
   dateCreated: 'Created:',
   dateLastUpdated: 'Updated:',
   userId: 'User ID:'
+
+
+  // api calls
+  apiCallNotValid: 'Invalid http request',
 }