+++ src/utils/customQuery/makeCustomQuery.js
@@ -1,15 +1,17 @@
-import axios from 'axios'
+// import axios from 'axios'
 import {
   NOTIFY_WARNING
 } from '../../constants/notifications'
 import showNotification from '../notifications/showNotification'
 import { GET_GRAPH_QUERY } from '../../constants/api'
 import store from '../../store'
+import httpCall from '../apiCalls/httpCall'
 
 /**
  * Get graph data from API
  * @param  {Object}   params
  * @param  {Function} params.setStoreState              setStoreState action
+ * @param  {Function} params.addToArray                 addToArray action
  * @param  {Function} params.t                          i18n translation function
  * @return {undefined}
  */
@@ -23,39 +25,36 @@ const makeCustomQuery = async ({
     user
   } = store.getState()
 
-  setStoreState('loading', true)
-
   setStoreState('customQueryFromLatestOutput', '')
   addToArray('customQueryStringHistory', customQueryString, { alwaysAdd: true })
 
-  try {
-    const response = await axios.post(GET_GRAPH_QUERY, {
+  const withAuth = !!user.token
+
+  const response = await httpCall({
+    setStoreState,
+    withAuth,
+    body: {
       query: customQueryString
-    }, {
-      headers: {
-        Authorization: user.token
-      }
-    })
+    },
+    route: GET_GRAPH_QUERY,
+    method: 'post',
+    t
+  })
 
-    setStoreState('loading', false)
+  const {
+    error,
+    data
+  } = response
 
-    if (response.status !== 200) {
-      return showNotification({
-        message: t('couldNotQueryGraph'),
-        type: NOTIFY_WARNING
-      })
-    }
-
-    setStoreState('customQueryFromLatestOutput', customQueryString)
-    setStoreState('customQueryOutput', response.data)
-  } catch (error) {
-    setStoreState('loading', false)
-
-    showNotification({
+  if (error) {
+    return showNotification({
       message: t('couldNotQueryGraph'),
       type: NOTIFY_WARNING
     })
   }
+
+  setStoreState('customQueryFromLatestOutput', customQueryString)
+  setStoreState('customQueryOutput', data)
 }
 
 export default makeCustomQuery