+++ src/utils/editOntology/setOntologyRestoreEdge.js
@@ -2,17 +2,23 @@ import store from '../../store'
 import addEdge from '../nodesEdgesUtils/addEdge'
 import getNode from '../nodesEdgesUtils/getNode'
 import setEdgeStylesByProperty from '../networkStyling/setEdgeStylesByProperty'
+import { POST_CREATE_EDGE } from '../../constants/api'
+import httpCall from '../apiCalls/httpCall'
+import showNotification from '../notifications/showNotification'
+import { NOTIFY_SUCCESS, NOTIFY_WARNING } from '../../constants/notifications'
 
 /**
  * ADd ontology edge
  * @param  {Object}         params
  * @param  {Function}       params.setStoreState              setStoreState action
  * @param  {Object}         params.selectedElement  Element properties with from,to,edge keys
+ * @param  {Function}       params.t                          i18n function
  * @return {undefined}
  */
-const setOntologyRestoreEdge = ({
+const setOntologyRestoreEdge = async ({
   setStoreState,
   selectedElement,
+  t
 }) => {
   const {
     objectPropertiesFromApi,
@@ -29,15 +35,52 @@ const setOntologyRestoreEdge = ({
   const newNodesEdges = JSON.parse(JSON.stringify(nodesEdges))
   const newEdgesPerNode = JSON.parse(JSON.stringify(edgesPerNode))
 
-  // remove selected elements from deleted connection
-  const newDeletedEdges = deletedEdges.filter((connection) => !selectedElement.includes(connection))
-  setStoreState('deletedEdges', newDeletedEdges)
+  const restoredEdges = {}
 
   // restore connections from graph
   if (selectedElement.length > 0) {
-    selectedElement.map((element) => {
+    for (let index = 0; index < selectedElement.length; index++) {
+      const oldId = selectedElement[index]
+
+      const body = newObjectPropertiesFromApiBackup[oldId] ? JSON.parse(JSON.stringify(newObjectPropertiesFromApiBackup[oldId])) : undefined
+
+      if (!body) return false
+
+      body.label = 'subclass'
+
+      const response = await httpCall({
+        setStoreState,
+        withAuth: true,
+        route: POST_CREATE_EDGE,
+        method: 'post',
+        body,
+        t
+      })
+
+      const {
+        error, data
+      } = response
+
+      const message = `${t('couldNotRestoreNode')}: ${oldId}`
+      if (error) {
+        showNotification({
+          message,
+          type: NOTIFY_WARNING
+        })
+        continue
+      }
+
+      if (!data || Object.keys(data).length !== 1) {
+        showNotification({
+          message,
+          type: NOTIFY_WARNING
+        })
+        continue
+      }
+
+      const { id } = data[Object.keys(data)[0]]
+
       // add to object properties
-      const id = element
       newObjectPropertiesFromApi[id] = newObjectPropertiesFromApiBackup[id]
 
       const {
@@ -82,14 +125,30 @@ const setOntologyRestoreEdge = ({
         })
       }
 
-      return true
-    })
+      restoredEdges[id] = oldId
+    }
   }
+
+  // remove selected elements from deleted connection
+  const newDeletedEdges = deletedEdges.filter((connection) => !restoredEdges[connection])
+  setStoreState('deletedEdges', newDeletedEdges)
 
   // add data
   setStoreState('nodesEdges', newNodesEdges)
   setStoreState('edgesPerNode', newEdgesPerNode)
   setStoreState('objectPropertiesFromApi', newObjectPropertiesFromApi)
+
+  const restoredEdgesOldIds = Object.keys(restoredEdges)
+
+  if (restoredEdgesOldIds.length > 0) {
+    const restoredEdgesIds = restoredEdgesOldIds.map((edgeId) => restoredEdgesOldIds[edgeId])
+
+    const message = `${t('edgesRestored')}: ${restoredEdgesIds.join(', ')}`
+    showNotification({
+      message,
+      type: NOTIFY_SUCCESS
+    })
+  }
 }
 
 export default setOntologyRestoreEdge