+++ CHANGELOG.md
@@ -18,10 +18,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [0.4.0]
 ### Added
+- ONTOLOGY-93 API endpoints to create/update/delete edges
+- ONTOLOGY-93 API endpoints to create/update/delete nodes
+- ONTOLOGY-93 Added httpCall function
 - ONTOLOGY-26 Structured search
 - ONTOLOGY-78 Login page with login endpoint
 
 ### Changed
+- ONTOLOGY-93 Refactored edge management components and utils
+- ONTOLOGY-93 Refactored node management components and utils
+- ONTOLOGY-93 Refactored functions with api calls to use the httpCall function
 
 ### Deprecated